split out api strings; install peer dependencies
