+++ api-constants.js
@@ -1,0 +1,12 @@
+const ALLOWED_ENVIRONMENTS = [
+  'test',
+  'esm',
+  'cjs',
+  'development-compat',
+  'development-modern',
+  'production-compat',
+  'production-modern',
+]
+module.exports = {
+  ALLOWED_ENVIRONMENTS,
+}
