+++ package.json
@@ -33,12 +33,20 @@
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
+    "@babel/eslint-parser": "^7.13.4",
     "@jest/globals": "^26.6.2",
     "@kensho-technologies/eslint-config": "^21.0.0",
     "@kensho-technologies/prettier-config": "^1.0.0",
     "@kensho-technologies/tsconfig": "^2.0.0",
     "@types/babel__core": "^7.1.12",
+    "@typescript-eslint/eslint-plugin": "^4.15.2",
+    "@typescript-eslint/parser": "^4.15.2",
     "eslint": "^7.16.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsdoc": "^32.2.0",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
     "typescript": "^4.1.3"
